+++ s2_fish/plot_solver.py
@@ -41,9 +41,9 @@ t_solve = time.time() - t0
 print "init =",t_init,", deposit =",t_deposit,", solve =",t_solve
 num_points = n
 ax = numpy.arrayrange(num_points)*size[0]/(num_points -1) - 0.5*size[0]
-aphi = numpy.zeros([num_points],numpy.Float)
-arho = numpy.zeros([num_points],numpy.Float)
-aexact = numpy.zeros([num_points],numpy.Float)
+aphi = numpy.zeros([num_points],numpy.float64)
+arho = numpy.zeros([num_points],numpy.float64)
+aexact = numpy.zeros([num_points],numpy.float64)
 index = 0;
 for i in range(0,len(ax)):
     x = ax[i]
